requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
torch
torchvision
tqdm
pyyaml
jinja2
pytest
ruff
